--- src/main/java/jamdoggie/betterbattletowers/util/ParticleHelper.java
@@ -1,25 +1,0 @@
-package jamdoggie.betterbattletowers.util;
-
-import net.minecraft.core.net.packet.PacketAddParticle;
-import net.minecraft.core.world.World;
-import net.minecraft.server.MinecraftServer;
-import net.minecraft.server.net.PlayerList;
-import turniplabs.halplibe.helper.EnvironmentHelper;
-
-//public class ParticleHelper {
-//	public static void spawnParticle(World world, String particleKey, double x, double y, double z, double motionX, double motionY, double motionZ, int data, double maxDistance) {
-//		if (EnvironmentHelper.isClientWorld()) return;
-//
-//		if (EnvironmentHelper.isServerEnvironment()) {
-//			PlayerList playerList = MinecraftServer.getInstance().playerList;
-//
-//			playerList.sendPacketToAllPlayersInDimension(
-//				new PacketAddParticle(particleKey, x, y, z, motionX, motionY, motionZ, data, maxDistance),
-//				world.dimension.id
-//			);
-//
-//			return;
-//		}
-//		world.spawnParticle(particleKey, x, y, z, motionX, motionY, motionZ, data, maxDistance);
-//	}
-//}
