+++ src/main/java/jamdoggie/betterbattletowers/util/TowerProperties.java
@@ -1,12 +1,10 @@
-package jamdoggie.betterbattletowers.entity;
+package jamdoggie.betterbattletowers.util;
 
-import jamdoggie.betterbattletowers.util.BlockData;
 import net.minecraft.core.WeightedRandomBag;
 import net.minecraft.core.world.biome.Biome;
 
 import java.util.*;
 
-import static jamdoggie.betterbattletowers.entity.TowerProperties.BlockPallet.pallet;
 import static net.minecraft.core.block.Blocks.*;
 import static net.minecraft.core.world.biome.Biomes.*;
 
@@ -64,16 +62,16 @@ public class TowerProperties {
 
 	public static class TowerProperty {
 		WeightedRandomBag<BlockData> towerDecorations;
-		int skinVariant;
-		private TowerProperty(WeightedRandomBag<BlockData> towerDecorations, int skinVariant){
+		String skinVariant;
+		private TowerProperty(WeightedRandomBag<BlockData> towerDecorations, String skinVariant){
 			this.skinVariant = skinVariant;
 			this.towerDecorations = towerDecorations;
 		}
 		public WeightedRandomBag<BlockData> getTowerDecorations() {return towerDecorations;}
-		public int getSkinVariant() {return skinVariant;}
+		public String getSkinVariant() {return skinVariant;}
 	}
 
-	public static TowerProperty register(Set<Biome> biomeList, WeightedRandomBag<BlockData> blockBag, int variant) {
+	public static TowerProperty register(Set<Biome> biomeList, WeightedRandomBag<BlockData> blockBag, String variant) {
 		TowerProperty tower = new TowerProperty(blockBag, variant);
 		for (Biome biome : biomeList) {
 			biomeToTowerProperties.computeIfAbsent(biome, key -> new HashSet<>()).add(tower);
@@ -115,21 +113,21 @@ public class TowerProperties {
 				OVERWORLD_MEADOW, OVERWORLD_BIRCH_FOREST
 			),
 			asBag(BlockPallet.pallet(BRICK_STONE_POLISHED.id(), 45.0f), BlockPallet.pallet(STONE_CARVED.id(), 45.0f), BlockPallet.pallet(STONE_POLISHED.id(), 10.0f))
-			, 0
+			, "normal"
 		);
 		BRICK_STONE_TOWER = register(asSet(
 				OVERWORLD_RAINFOREST, OVERWORLD_SEASONAL_FOREST, OVERWORLD_FOREST, OVERWORLD_GRASSLANDS, OVERWORLD_BOREAL_FOREST, OVERWORLD_PLAINS,
 				OVERWORLD_MEADOW, OVERWORLD_BIRCH_FOREST
 			),
 			asBag(BlockPallet.pallet(BRICK_STONE.id(), 50.0f), BlockPallet.pallet(STONE.id(), 50.0f))
-			, 0
+			, "normal"
 		);
 		DILAPIDATED_STONE_TOWER = register(asSet(
 				OVERWORLD_RAINFOREST, OVERWORLD_SEASONAL_FOREST, OVERWORLD_FOREST, OVERWORLD_GRASSLANDS, OVERWORLD_BOREAL_FOREST, OVERWORLD_PLAINS,
 				OVERWORLD_MEADOW, OVERWORLD_BIRCH_FOREST
 			),
 			asBag(BlockPallet.pallet(COBBLE_STONE.id(), 50.0f), BlockPallet.pallet(COBBLE_STONE_MOSSY.id(), 50.0f))
-			, 0
+			, "normal"
 		);
 
 		RETRO_TOWER = register(asSet(
@@ -137,49 +135,49 @@ public class TowerProperties {
 				OVERWORLD_PLAINS, OVERWORLD_MEADOW, OVERWORLD_BIRCH_FOREST
 			),
 			asBag(BlockPallet.pallet(COBBLE_STONE.id(), 100.0f))
-			, 0
+			, "normal"
 		);
 
 		DESERT_TOWER = register(asSet(OVERWORLD_DESERT, OVERWORLD_CAATINGA),
 			asBag(BlockPallet.pallet(BRICK_SANDSTONE.id(), 50.0f), BlockPallet.pallet(SANDSTONE.id(), 50.0f))
-			, 1
+			, "hot"
 		);
 		DILAPIDATED_DESERT_TOWER = register(asSet(OVERWORLD_DESERT, OVERWORLD_CAATINGA),
 			asBag(BlockPallet.pallet(BRICK_SANDSTONE.id(), 30.0f), BlockPallet.pallet(SANDSTONE.id(), 40.0f), BlockPallet.pallet(SAND.id(), 30.0f))
-			, 1
+			, "hot"
 		);
 
 		GRANITE_TOWER = register(asSet(
 				OVERWORLD_OUTBACK, OVERWORLD_SHRUBLAND, OVERWORLD_DESERT, OVERWORLD_CAATINGA, OVERWORLD_CAATINGA_PLAINS
 			),
 			asBag(BlockPallet.pallet(GRANITE_CARVED.id(), 50.0f), BlockPallet.pallet(GRANITE_POLISHED.id(), 50.0f))
-			, 2
+			, "hot"
 		);
 		DILAPIDATED_GRANITE_TOWER = register(asSet(
 				OVERWORLD_OUTBACK, OVERWORLD_SHRUBLAND, OVERWORLD_DESERT, OVERWORLD_CAATINGA, OVERWORLD_CAATINGA_PLAINS
 			),
 			asBag(BlockPallet.pallet(BRICK_GRANITE.id(), 45.0f), BlockPallet.pallet(GRANITE.id(), 45.0f), BlockPallet.pallet(COBBLE_GRANITE.id(), 10.0f))
-			, 2
+			, "hot"
 		);
 
 		LIMESTONE_TOWER = register(asSet(
 				OVERWORLD_FOREST, OVERWORLD_GRASSLANDS, OVERWORLD_SHRUBLAND, OVERWORLD_PLAINS, OVERWORLD_MEADOW
 			),
 			asBag(BlockPallet.pallet(LIMESTONE_POLISHED.id(), 50.0f), BlockPallet.pallet(LIMESTONE_CARVED.id(), 50.0f))
-			, 3
+			, "normal"
 		);
 		DILAPIDATED_LIMESTONE_TOWER = register(asSet(
 				OVERWORLD_FOREST, OVERWORLD_GRASSLANDS, OVERWORLD_SHRUBLAND, OVERWORLD_PLAINS, OVERWORLD_MEADOW
 			),
 			asBag(BlockPallet.pallet(BRICK_LIMESTONE.id(), 50.0f), BlockPallet.pallet(COBBLE_LIMESTONE.id(), 50.0f))
-			, 3
+			, "normal"
 		);
 		BASALT_TOWER = register(asSet(
 				OVERWORLD_SEASONAL_FOREST, OVERWORLD_FOREST, OVERWORLD_GRASSLANDS, OVERWORLD_SHRUBLAND, OVERWORLD_BOREAL_FOREST,
 				OVERWORLD_PLAINS, OVERWORLD_MEADOW, OVERWORLD_BIRCH_FOREST, OVERWORLD_HELL
 			),
 			asBag(BlockPallet.pallet(BASALT_POLISHED.id(), 50.0f), BlockPallet.pallet(BASALT_CARVED.id(), 50.0f))
-			, 4
+			, "hot"
 		);
 
 		DILAPIDATED_BASALT_TOWER = register(asSet(
@@ -187,38 +185,38 @@ public class TowerProperties {
 				OVERWORLD_PLAINS, OVERWORLD_MEADOW, OVERWORLD_BIRCH_FOREST, OVERWORLD_HELL
 			),
 			asBag(BlockPallet.pallet(BASALT_POLISHED.id(), 45.0f), BlockPallet.pallet(BASALT_CARVED.id(), 45.0f), BlockPallet.pallet(BASALT.id(), 10.0f))
-			, 4
+			, "hot"
 		);
 
 		OVERGROWN_TOWER = register(asSet(
 				OVERWORLD_RAINFOREST, OVERWORLD_SWAMPLAND, OVERWORLD_FOREST, OVERWORLD_BIRCH_FOREST, OVERWORLD_SWAMPLAND_MUDDY
 			),
 			asBag(BlockPallet.pallet(BRICK_STONE_POLISHED_MOSSY.id(), 70.0f), BlockPallet.pallet(BRICK_STONE_POLISHED.id(), 15.0f), BlockPallet.pallet(STONE_CARVED.id(), 15.0f))
-			, 5
+			, "normal"
 		);
 		DILAPIDATED_OVERGROWN_TOWER = register(asSet(
 				OVERWORLD_RAINFOREST, OVERWORLD_SWAMPLAND, OVERWORLD_FOREST, OVERWORLD_BIRCH_FOREST, OVERWORLD_SWAMPLAND_MUDDY
 			),
 			asBag(BlockPallet.pallet(BRICK_STONE_POLISHED_MOSSY.id(), 50.0f), BlockPallet.pallet(MOSS_STONE.id(), 40.0f), BlockPallet.pallet(MOSS_LIMESTONE.id(), 10.0f))
-			, 5
+			, "normal"
 		);
 		FROZEN_TOWER = register(asSet(
 				OVERWORLD_TAIGA, OVERWORLD_BOREAL_FOREST, OVERWORLD_GLACIER, OVERWORLD_TUNDRA
 			),
 			asBag(BlockPallet.pallet(BRICK_PERMAFROST.id(), 45.0f), BlockPallet.pallet(PERMAFROST_CARVED.id(), 45.0f), BlockPallet.pallet(PERMAFROST_POLISHED.id(), 10.0f))
-			, 6
+			, "cold"
 		);
 		DILAPIDATED_FROZEN_TOWER = register(asSet(
 				OVERWORLD_TAIGA, OVERWORLD_BOREAL_FOREST, OVERWORLD_GLACIER, OVERWORLD_TUNDRA
 			),
 			asBag(BlockPallet.pallet(PERMAFROST.id(), 45.0f), BlockPallet.pallet(COBBLE_PERMAFROST.id(), 45.0f), BlockPallet.pallet(PERMAICE.id(), 10.0f))
-			, 6
+			, "cold"
 		);
 		SNOW_TOWER = register(asSet(
 				OVERWORLD_GLACIER, OVERWORLD_TUNDRA
 			),
 			asBag(BlockPallet.pallet(BLOCK_SNOW.id(), 100.0f))
-			, 6
+			, "cold"
 		);
 
 		PARADISE_TOWER = register(asSet(
@@ -226,27 +224,27 @@ public class TowerProperties {
 				OVERWORLD_BOREAL_FOREST, OVERWORLD_PLAINS, OVERWORLD_MEADOW, OVERWORLD_BIRCH_FOREST, OVERWORLD_HELL
 			),
 			asBag(BlockPallet.pallet(BRICK_MARBLE.id(), 100.0f))
-			, 7
+			, "cold"
 		);
 		HELLFIRE_TOWER = register(asSet(OVERWORLD_HELL),
 			asBag(BlockPallet.pallet(NETHERRACK_POLISHED.id(), 50.0f), BlockPallet.pallet(NETHERRACK_CARVED.id(), 50.0f))
-			, 8
+			, "hot"
 		);
 		OBSIDIAN_TOWER = register(asSet(OVERWORLD_HELL),
 			asBag(BlockPallet.pallet(OBSIDIAN.id(), 100.0f))
-			, 8
+			, "hot"
 		);
 		SLAVERS_TOWER = register(asSet(OVERWORLD_HELL),
 			asBag(BlockPallet.pallet(BRICK_NETHERRACK.id(), 100.0f))
-			, 8
+			, "hot"
 		);
 		DILAPIDATED_NETHERRACK_TOWER = register(asSet(OVERWORLD_HELL),
 			asBag(BlockPallet.pallet(NETHERRACK.id(), 49.94f), BlockPallet.pallet(COBBLE_NETHERRACK.id(), 49.94f), BlockPallet.pallet(ORE_NETHERCOAL_NETHERRACK.id(), 0.02f))
-			, 8
+			, "hot"
 		);
 		SLATE_TOWER = register(asSet(OVERWORLD_HELL),
 			asBag(BlockPallet.pallet(SLATE_POLISHED.id(), 50.0f), BlockPallet.pallet(SLATE.id(), 50.0f))
-			, 8
+			, "hot"
 		);
 
 	}
